chore: spreadsheet view context
